Wrap some Invoke-RestMethod in a retry loop
Sometimes I have network errors when talking to my ElasticSearch. This is
really inconvenient when I'm loading biiiiig scans and have to restart an
operation which might take up to an hour.

According to the powershell docs `MaximumRetryCount` only protects from http
errors (or so it seems), not network errors. This wraps Invoke-RestMethod in
a try/catch loop so that the failing request is retried.